Use SignalHandler instead of sigmux in SamplingProfiler
Summary:
This diff uses the new facilities in `SignalHandler` to replace `sigmux` usage and always take over the signal flow before any previously installed handler.

# A note on chaining to the previous handler
Because there's no way to use `sigaction` safely to revert to a previous handler, once installed our handlers are never removed but merely disabled.

When disabled, SIGSEGV/SIGBUS handler delegates to the previous handler but the SIGPROF handler *does not* delegate to the previous handler. This is because we can't tell if the previously installed handler (e.g. debuggerd's handler) would take down the process as part of its execution. The default disposition for SIGPROF is Term, so we can't afford to let such a handler run.

Why can't we let it run when our SignalHandler is disabled, so we catch other bad sources of SIGPROF? Because we can't know if the SIGPROF being delivered was queued due to our profiler. In the case of `signal queued`, `stopProfling`, `signal handler executes`, we need to still ignore the signal or else we'd take down the process.

Reviewed By: yukonfb

Differential Revision: D23248678

fbshipit-source-id: 0f38e7af477e5eb41ad062ab8a379173627ae048