Extract a base Buffer class that abstracts backing store
Summary:
We want to represent the ring buffer and its backing store in an object that owns both.
Unfortunately, LockFreeRingBufferHolder only cares about owning the memory, not the underlying file or header/prefix, so it's not a very suitable candidate for this job.

Therefore, we introduce `Buffer` (currently in the `mmapbuf` namespace but this will change). In this diff, it owns the underlying file, as well as the header/prefix. In future diffs, it will own the TraceBuffer instance as well.

In doing this, we mimic the semantics from the existing `MmapBufferManager` but with a few catches:
* destroying a file backed `Buffer` object unlinks the underlying file.
* we always allocate a prefix, even for anonymous memory buffers, to simplify handling.

Reviewed By: aandreyeu

Differential Revision: D23658662

fbshipit-source-id: e3a866fb99214f356211e07960b804fb97358b27