Enable prefab native Android distribution (#43)
Summary:
This enables the long-awaited "prefab" mode which is a proper binary and header distribution for pre-built binaries. More on that here: https://google.github.io/prefab/.

This will likely break backwards-compatibility with packages not using prefab, so we will want to at least bump the minor version for the next release. In case that this causes too much trouble, we can publish a separate non-prefab distribution, but we will tackle that reactively.

Pull Request resolved: https://github.com/facebookincubator/fbjni/pull/43

Test Plan:
./gradlew assemble

`unzip -l build/outputs/aar/fbjni-release.aar` shows we've got headers (well, actually all source) plus binaries in a `prefab/` folder.

Checked that this builds with Flipper by publishing it locally and depending on it. PR for this will follow.

Reviewed By: fabiomassimo

Differential Revision: D26517673

Pulled By: passy

fbshipit-source-id: a05087f98df2116793ec37b1ffeb358684ae1dfe