Add release GitHub workflow
Summary:
This will do the whole release flow (build, upload, close repository)
whenever either a tag is created or the workflow is manually triggered.

Blindly written. Can't really test this until all the other diffs have landed.
This is effectively the same setup as for Flipper, so it shouldn't cause a massive
amount of problems, but we will see.

The secrets are already all set up.

Reviewed By: fabiomassimo

Differential Revision: D26546834

fbshipit-source-id: 238521317207e6b61f3c9a129474eae39967f498