Implement TraceWriter::dump for secondary buffers
Summary:
Secondary buffers (i.e. other `Buffer`s in `TraceContext.buffers` that are not
`TraceContext.mainBuffer) do not contain trace lifecycle events (TRACE_START
etc). Therefore, we need a new way to produce files for them.

This diff introduces `TraceWriter::dump` to dump all entries in the buffer into
a file and integrates it with `LoggerWorkerThread`. It also allows
`NativeTraceWriter` to have `null` callbacks, as callbacks are not really
meaningful for `dump`.

One unfortunate complication is that we need to delay trace lifecycle callbacks
(`onTraceWriteEnd`/`Abort`) to until after the secondary buffers have been
dumped. To achieve this, we implement `CachingNativeTraceWriterCallbacks`,

As a side effect, if a latency trace has extra buffers, they would also get
dumped at the end of the trace (but not followed throughout the latency trace).

Reviewed By: aandreyeu

Differential Revision: D26957883

fbshipit-source-id: 88dce79d76a3651590f4e29d1add4b954f8fedb1