Allow processing of onTraceWriteException before onTraceStartAsync
Summary:
This is a fun little race condition that we ended up making a bit more common.

If an exception is thrown on the background thread before onTraceAsync has had
a chance to run, we could trigger a crash by breaking the lookup of the `Trace`
object.

Technically, this has existed for ages (e.g., [sample before the recent changes](https://fburl.com/scuba/errorreporting_facebook_android_crashes/b1uszev8))
but that doesn't mean we can't fix it. :)

Reviewed By: aandreyeu

Differential Revision: D27194803

fbshipit-source-id: 24f0afd9086e6e7243cdb11792fa79186db99dfe