Fixes for -Wextra-semi,-Wextra-semi-stmt
Summary: The flags are being introduced to higher level modules that depend on these headers. Although I admit these specific flags are more aesthetic than functional, I still would like to keep them around, and it's also normal for lower level libraries to be subject to a higher bar in terms of quality and readability.

Reviewed By: mhorowitz

Differential Revision: D41164579

fbshipit-source-id: 03e8a4c23c5ba83e791b0b442bc63c9d88cc6c98